Removed "top parents"
I removed the "top parents" parameter entirely because I determined it wasn't necessary anymore. It is a little too complicated to explain in this box right now why exactly I felt it necessary in the first place, but it did have to do with the fact that I initially conceived of the "Genre" class without a "subgenres" parameter, meaning that there was no quick way to access a Genre's children. "top_parents" was eventually going to lead to a fast method, but I determined "subgenres" would lead to much less of a headache and be more convenient overall.