Working with CSV
Now I decided to go into the larger Genre Tree Program and write a "with open as file" algorithm that can read each row in a csv file and add a genre based on the row's information.

I tested it out using a sample of 100 genres in my 'genres2' csv file.