Test Case: Print All Genres
I made the "with open as file" algorithm into a function.

I also wrote an algorithm (not yet a function) that would return a list of all the genres with their .__str__() representations and after some tinkering (and also making it so the .__str__ returned string parents and not object parents), I got it. This successfully returned all 100 genres I put in, and in the expected order.