Testing the Program with all 870 Genres
I finally decided to test out the program with all 870 genres (note: It is less than 870 because there are many rows that have a repeat genre, plus there's "Top-level" and "No Genre" which are not actual genres and just for utility).

I put in "encoding='utf-8'" into the csv_extract function to account for an error that occurred while first testing.