Back Main Multiple
Made the long-awaited "back_main_multiple" function, which can accept a list of genres (separated by “; ”) and return the sheet name of all those genres followed by the parents of all those genres, if said parents’ names are not already contained within an earlier genre name.

As part of making this function, I also adjusted the "back_main" function so that there is a 'str' mode and an 'obj' mode (the former makes the function return a string, and the latter makes the function return a list of objects). This was very helpful in making the back_main_multiple function work.

I tested it out with 'Chillwave; Jangle Pop; Country Rock' and it correctly returned 'Chillwave; Jangle Pop; C0untry Rock; Indietronica; Electronic; Pop Rock; Alt Rock' ("Country Rock" changing to "C0untry Rock" is intentional)