Solution to Recursive Issue
I will admit this solution is relatively crude, but it works. This essentially circumvents the object list complication by simply taking the back_all_genres_n string (a version of the back_all_genres string that uses the pure name, not the sheet name), removing the "; None" via .replace(), using .split() with a delimiter of "; " to make a list, and then making a list of objects from this list using the dictionary key method in back_main_multiple. It is too much to explain here and is better documented elsewhere, but that is my attempt at explaining the changes here.