Fixing the Language Variable Problem
I did a lot of tinkering around in the Inspect Element console of a few RYM pages and with that, I was able to figure out how to check if there's a Language field and then how to find the relevant Language field. I was still able to use the "Copy JS path" (see the var lang_field) because the nth child included a number that corresponded to the lang_row_num, plus one. It's a little complicated to explain on here, but long story short, I fixed the problem of language variables sometimes not showing up due to varying numbers of rows in each page. I looked at various different types of pages for this