Major Overhaul of back_all
While attempting to add new genres en masse, I found a flaw in back_all that unraveled more flaws with it to the point where I had to duplicate my file in order to go in and reconstruct the back_all function. It wasn't from scratch, but I did have to significantly change it.

I already document my processes in a separate log, so just look there. It's a little too much to explain here right now.

Also genres3.2 got modified to add "Folk" as a parent for "Delta Blues", so the name was changed to genres3.25