Blood Functions
The main theme of this session was making a function that could check two genres and see if they have a parent-child relationship, which is important for being able to take a back genre list and "reverse engineer" it back to a front genre list. I actually used these functions for a personal spreadsheet I have. I inputted a csv file with 2000 rows and it managed to work as intended.

My process is all better documented in a separate log, anyway, but that's the gist of it