"gs" dictionary for Sheet Names
Made a separate dictionary called ‘gs’ that’s just for the genres with the sheet names, so now the genres with sheet names can be referred to using their pure version OR their sheet version as keys.

Just to be clear, the genres with sheet names still go in the g dictionary, they just additionally go in the 'gs' dictionary as well with a different key. Same object, two places.