Accounted for artists whose names start with "The"
Made it so that if an artist’s name starts with “The” (e.g. “The Beach Boys”), the “The ” part of their name is deleted when entered into the table, so the table will only have “Beach Boys”. However, whenever using the clipboard functions, it still will copy “The Beach Boys” to the clipboard. This is intentional because the removal of “The” is only relevant to the spreadsheet cell